Split generic and Linux broker transports
Keep peer-visible contracts in the protocol crate, move runtime channel and shared-memory abstractions into the no_std transport crate, and isolate memfd, futex, Unix framing, and both concrete endpoints in a Linux-userland transport crate.


Copilot-Session: b5a1a347-37a8-4246-8bbc-306590921475
