+++ Cargo.toml
@@ -6,8 +6,8 @@ members = [
     "litebox_broker_core",
     "litebox_broker_protocol",
     "litebox_broker_host",
-    "litebox_broker_platform_linux_userland",
     "litebox_broker_transport",
+    "litebox_broker_transport_linux_userland",
     "litebox_broker_userland",
     "litebox_common_linux",
     "litebox_common_windows",
@@ -42,8 +42,8 @@ default-members = [
     "litebox_broker_core",
     "litebox_broker_protocol",
     "litebox_broker_host",
-    "litebox_broker_platform_linux_userland",
     "litebox_broker_transport",
+    "litebox_broker_transport_linux_userland",
     "litebox_broker_userland",
     "litebox_common_linux",
     "litebox_common_windows",
