+++ dev_tests/src/ratchet.rs
@@ -72,7 +72,7 @@ fn ratchet_maybe_uninit() -> Result<()> {
         &[
             ("dev_tests/", 1),
             ("litebox/", 1),
-            ("litebox_broker_transport/", 3),
+            ("litebox_broker_transport_linux_userland/", 3),
             ("litebox_platform_linux_userland/", 2),
         ],
         |file| {
