+++ litebox/src/pipes.rs
@@ -928,7 +928,6 @@ mod tests {
 
     use alloc::sync::Arc;
     use litebox_broker_local::BrokerLocal;
-    use litebox_broker_protocol::channel::{LocalCallChannel, LocalSetupChannel};
     use litebox_broker_protocol::error::ErrorCode;
     use litebox_broker_protocol::message::{
         BrokerHandshakeRequest, BrokerHandshakeResponse, BrokerNotification, BrokerOperation,
@@ -937,6 +936,7 @@ mod tests {
     use litebox_broker_protocol::pipe::CreatePipeResponse;
     use litebox_broker_protocol::readiness::ReadinessFlags;
     use litebox_broker_protocol::{BROKER_PROTOCOL_VERSION, ObjectHandle};
+    use litebox_broker_transport::channel::{LocalCallChannel, LocalSetupChannel};
 
     use crate::{
         event::{Events, observer::Observer, wait::WaitState},
@@ -1117,16 +1117,16 @@ mod tests {
     #[derive(Clone, Copy)]
     struct NoopSharedMemory;
 
-    impl litebox_broker_protocol::shared_memory::SharedMemory for NoopSharedMemory {
+    impl litebox_broker_transport::shared_memory::SharedMemory for NoopSharedMemory {
         fn len(&self) -> usize {
-            litebox_broker_protocol::shared_memory::SHARED_BUFFER_POOL_SIZE
+            litebox_broker_protocol::shared_buffer::SHARED_BUFFER_POOL_SIZE
         }
 
         fn read(
             &self,
             _offset: usize,
             destination: &mut [u8],
-        ) -> core::result::Result<(), litebox_broker_protocol::shared_memory::SharedMemoryError>
+        ) -> core::result::Result<(), litebox_broker_transport::shared_memory::SharedMemoryError>
         {
             destination.fill(0);
             Ok(())
@@ -1136,7 +1136,7 @@ mod tests {
             &self,
             _offset: usize,
             _source: &[u8],
-        ) -> core::result::Result<(), litebox_broker_protocol::shared_memory::SharedMemoryError>
+        ) -> core::result::Result<(), litebox_broker_transport::shared_memory::SharedMemoryError>
         {
             Ok(())
         }
