+++ litebox_broker_transport/Cargo.toml
@@ -3,23 +3,9 @@ name = "litebox_broker_transport"
 version = "0.1.0"
 edition = "2024"
 
-[features]
-std = []
-linux-userland = [
-    "std",
-    "dep:libc",
-    "dep:rustix",
-    "rustix/fs",
-    "rustix/net",
-    "rustix/thread",
-]
-
 [dependencies]
 litebox_broker_protocol = { path = "../litebox_broker_protocol", version = "0.1.0" }
-
-[target.'cfg(target_os = "linux")'.dependencies]
-libc = { version = "0.2.177", default-features = false, optional = true }
-rustix = { version = "1.1.2", default-features = false, features = ["std"], optional = true }
+thiserror = { version = "2.0.6", default-features = false }
 
 [lints]
 workspace = true
