+++ litebox_broker_transport_linux_userland/Cargo.toml
@@ -1,12 +1,15 @@
 [package]
-name = "litebox_broker_local"
+name = "litebox_broker_transport_linux_userland"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
 litebox_broker_protocol = { path = "../litebox_broker_protocol", version = "0.1.0" }
 litebox_broker_transport = { path = "../litebox_broker_transport", version = "0.1.0" }
-thiserror = { version = "2.0.6", default-features = false }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+libc = { version = "0.2.177", default-features = false }
+rustix = { version = "1.1.2", default-features = false, features = ["std", "fs", "net", "thread"] }
 
 [lints]
 workspace = true
