+++ litebox_runner_linux_userland/Cargo.toml
@@ -10,7 +10,8 @@ libc = { version = "0.2.169", default-features = false }
 litebox = { version = "0.1.0", path = "../litebox" }
 litebox_broker_local = { version = "0.1.0", path = "../litebox_broker_local" }
 litebox_broker_protocol = { version = "0.1.0", path = "../litebox_broker_protocol" }
-litebox_broker_transport = { version = "0.1.0", path = "../litebox_broker_transport", features = ["linux-userland"] }
+litebox_broker_transport = { version = "0.1.0", path = "../litebox_broker_transport" }
+litebox_broker_transport_linux_userland = { version = "0.1.0", path = "../litebox_broker_transport_linux_userland" }
 litebox_common_linux = { version = "0.1.0", path = "../litebox_common_linux" }
 litebox_platform_linux_userland = { version = "0.1.0", path = "../litebox_platform_linux_userland" }
 litebox_shim_linux = { version = "0.1.0", path = "../litebox_shim_linux" }
@@ -25,7 +26,6 @@ walkdir = "2.0"
 glob = "0.3"
 litebox_broker_core = { version = "0.1.0", path = "../litebox_broker_core" }
 litebox_broker_host = { version = "0.1.0", path = "../litebox_broker_host" }
-litebox_broker_platform_linux_userland = { version = "0.1.0", path = "../litebox_broker_platform_linux_userland" }
 
 [features]
 lock_tracing = ["litebox/lock_tracing"]
