+++ litebox_runner_linux_userland/tests/run.rs
@@ -362,17 +362,18 @@ fn spawn_test_broker(
                     .accept()
                     .expect("failed to accept broker local control connection");
                 let shared_memory =
-                    litebox_broker_transport::shared_memory::MemfdSharedMemory::create(
-                        litebox_broker_protocol::shared_memory::SHARED_BUFFER_POOL_SIZE,
+                    litebox_broker_transport_linux_userland::shared_memory::MemfdSharedMemory::create(
+                        litebox_broker_protocol::shared_buffer::SHARED_BUFFER_POOL_SIZE,
                     )
                     .expect("failed to create broker test shared memory");
-                let shared_buffers = litebox_broker_protocol::shared_memory::SharedBufferPool::new(
-                    shared_memory,
-                    litebox_broker_protocol::shared_memory::SHARED_BUFFER_LAYOUT,
-                )
-                .expect("failed to attach broker test shared-buffer layout");
+                let shared_buffers =
+                    litebox_broker_transport::shared_memory::SharedBufferPool::new(
+                        shared_memory,
+                        litebox_broker_protocol::shared_buffer::SHARED_BUFFER_LAYOUT,
+                    )
+                    .expect("failed to attach broker test shared-buffer layout");
                 let control_memory =
-                    litebox_broker_transport::shared_memory::MemfdSharedMemory::create(
+                    litebox_broker_transport_linux_userland::shared_memory::MemfdSharedMemory::create(
                         litebox_broker_transport::control_ring::CONTROL_RING_MEMORY_SIZE,
                     )
                     .expect("failed to create broker test control ring");
@@ -386,7 +387,7 @@ fn spawn_test_broker(
                     .set_write_timeout(Some(BROKER_HELPER_TIMEOUT))
                     .expect("failed to configure broker test write timeout");
                 let mut channel =
-                    litebox_broker_platform_linux_userland::unix_socket::UnixStreamHostSetupChannel::from_host_guaranteed(
+                    litebox_broker_transport_linux_userland::unix_socket::UnixStreamHostSetupChannel::from_host_guaranteed(
                         control_stream,
                         std::time::Instant::now() + BROKER_HELPER_TIMEOUT,
                     );
@@ -410,7 +411,7 @@ fn spawn_test_broker(
                         .recv_request()
                         .expect("failed to receive broker test request")
                     {
-                        litebox_broker_protocol::channel::HostReceive::Message(request) => {
+                        litebox_broker_transport::channel::HostReceive::Message(request) => {
                             if matches!(
                                 &request.operation,
                                 litebox_broker_protocol::message::BrokerOperation::CloseObject(_)
@@ -423,10 +424,10 @@ fn spawn_test_broker(
                                 })
                                 .expect("failed to execute broker test request");
                         }
-                        litebox_broker_protocol::channel::HostReceive::PeerClosed => {
+                        litebox_broker_transport::channel::HostReceive::PeerClosed => {
                             break litebox_broker_host::ConnectionTermination::PeerClosed;
                         }
-                        litebox_broker_protocol::channel::HostReceive::ProtocolViolation => {
+                        litebox_broker_transport::channel::HostReceive::ProtocolViolation => {
                             break litebox_broker_host::ConnectionTermination::ProtocolViolation;
                         }
                     }
